-Aggiunti tutti i metodi necessari ad Evento.java -Aggiunti tutti i metodi per la gestione eventi necessari ad ControlloreUtente -Aggiunti tutti i metodi necessari per la gestione eventi ad ControlloreLocale -Aggiunti tab eventi e news nel dettaglio locale -Creato EventBean.java -Aggiunto form per aggiungere nuovo evento(tasto aggiungi non funzionante...) TODO: -stesso lavoro per le news -query di ricerca per gruppi di locali -funzione che cancelli i vecchi eventi

+++ trunk/branchJSF/luncharound/luncharound-ejb/src/java/luncharoundpkg/EventoFacadeLocal.java
@@ -32,6 +32,6 @@ public interface EventoFacadeLocal {
     
     public List<Evento> findByLocale(long idLocale );
     
-    public List<Evento> findByLocali(List<Long> idLocali );
+    public List<Evento> findByLocali(List<Integer> idLocali );
     
 }
